+++ toteatds/src/commonMain/kotlin/com/toteat/toteatds/components/topbar/LoginTopBar.kt
@@ -43,9 +43,7 @@ fun LoginTopBar(
                     }
             )
         },
-        rightComponent = {
-            CustomerServiceIcon(onIconClick = onCustomerServiceButtonClick)
-        },
+        rightComponent = {},
         testTag = testTag
     )
 }